Revert "Restructure"
