+++ .gitignore
@@ -2,8 +2,6 @@
 *.png
 !champion_priors.csv
 
-.vscode/
-
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
