+++ champ_rec.bat
@@ -1,7 +1,5 @@
 @echo off
-call "%~dp0.venv\Scripts\activate.bat"
-python "%~dp0champ_rec.py"
-
+"D:\Nedlastninger\app\SKADZALGORITM\champ_rec\.venv\Scripts\python.exe" "D:\Nedlastninger\app\SKADZALGORITM\champ_rec\champ_rec.py"
 pause
