--- core/repo/__init__.py
@@ -1,4 +1,0 @@
-from .matchup_repository import MatchupRepository
-from .priors_repository import PriorsRepository
-
-__all__ = ["MatchupRepository", "PriorsRepository"]
